Renamed /routes to /endpoints
