+++ pyproject.toml
@@ -14,7 +14,6 @@ asyncpg = "^0.23.0"
 gunicorn = "^20.1.0"
 starlette = "^0.14.2"
 loguru = "^0.5.3"
-aiohttp = "^3.7.4"
 
 [tool.poetry.dev-dependencies]
 pytest = "^6.2"
